Moves some test functions
